added exact_cover
