+++ qiskit_optimization/applications/ising/exact_cover.py
@@ -10,113 +10,43 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-""" exact cover """
+"""
+Convert vertex cover instances into Pauli list
+Deal with Gset format. See https://web.stanford.edu/~yyye/yyye/Gset/
+"""
+import copy
 
-from typing import Tuple, List
-import logging
-
+import networkx as nx
 import numpy as np
+from docplex.mp.model import Model
 
-from qiskit.quantum_info import Pauli
-from qiskit.opflow import PauliSumOp
-
-logger = logging.getLogger(__name__)
-
-
-def get_operator(list_of_subsets: List) -> Tuple[PauliSumOp, float]:
-    """
-    Construct the Hamiltonian for the exact solver problem.
-
-    Note:
-        | Assumption: the union of the subsets contains all the elements to cover.
-        | The Hamiltonian is:
-        |   sum_{each element e}{(1-sum_{every subset_i that contains e}{Xi})^2},
-        |   where Xi (Xi=1 or 0) means whether should include the subset i.
-
-    Args:
-        list_of_subsets: list of lists (i.e., subsets)
-
-    Returns:
-        operator for the Hamiltonian, a constant shift for the obj function.
-    """
-    # pylint: disable=invalid-name
-    n = len(list_of_subsets)
-
-    U = []
-    for sub in list_of_subsets:
-        U.extend(sub)
-    U = np.unique(U)  # U is the universe
-
-    shift = 0.
-    pauli_list = []
-
-    for e in U:
-        # pylint: disable=simplifiable-if-expression
-        cond = [True if e in sub else False for sub in list_of_subsets]
-        indices_has_e = np.arange(n)[cond]
-        num_has_e = len(indices_has_e)
-        Y = 1 - 0.5 * num_has_e
-        shift += Y * Y
-
-        for i in indices_has_e:
-            for j in indices_has_e:
-                if i != j:
-                    w_p = np.zeros(n)
-                    v_p = np.zeros(n)
-                    v_p[i] = 1
-                    v_p[j] = 1
-                    pauli_list.append([0.25, Pauli((v_p, w_p))])
-                else:
-                    shift += 0.25
-
-        for i in indices_has_e:
-            w_p = np.zeros(n)
-            v_p = np.zeros(n)
-            v_p[i] = 1
-            pauli_list.append([-Y, Pauli((v_p, w_p))])
-
-    opflow_list = [(pauli[1].to_label(), pauli[0]) for pauli in pauli_list]
-    return PauliSumOp.from_list(opflow_list), shift
+from .base_application import BaseApplication
+from qiskit_optimization.problems.quadratic_program import QuadraticProgram
 
 
-def get_solution(x):
-    """
-    Args:
-        x (numpy.ndarray) : binary string as numpy array.
+class ExactCover(BaseApplication):
 
-    Returns:
-        numpy.ndarray: graph solution as binary numpy array.
-    """
-    return 1 - x
+    def __init__(self, subsets):
+        self._subsets = copy.deepcopy(subsets)
+        self._set = []
+        for sub in self._subsets:
+            self._set.extend(sub)
+        self._set = np.unique(self._set)
 
+    def to_quadratic_program(self):
+        mdl = Model(name='exact_cover')
+        x = {i: mdl.binary_var(name='x_{0}'.format(i)) for i in range(len(self._subsets))}
+        mdl.minimize(mdl.sum(x[i] for i in x))
+        for e in self._set:
+            mdl.add_constraint(mdl.sum(x[i] for i, sub in enumerate(self._subsets)
+                                       if e in sub) == 1)
+        qp = QuadraticProgram()
+        qp.from_docplex(mdl)
+        return qp
 
-def check_solution_satisfiability(sol, list_of_subsets):
-    """ check solution satisfiability """
-    # pylint: disable=invalid-name
-    n = len(list_of_subsets)
-    U = []
-    for sub in list_of_subsets:
-        U.extend(sub)
-    U = np.unique(U)  # U is the universe
-
-    U2 = []
-    selected_subsets = []
-    for i in range(n):
-        if sol[i] == 1:
-            selected_subsets.append(list_of_subsets[i])
-            U2.extend(list_of_subsets[i])
-
-    U2 = np.unique(U2)
-    if set(U) != set(U2):
-        return False
-
-    tmplen = len(selected_subsets)
-    for i in range(tmplen):
-        for j in range(i):
-            L = selected_subsets[i]
-            R = selected_subsets[j]
-
-            if set(L) & set(R):  # should be empty
-                return False
-
-    return True
+    def interpret(self, result):
+        sub = []
+        for i, value in enumerate(result.x):
+            if value:
+                sub.append(self._subsets[i])
+        return sub
